# backend/docs/swagger.json
{
    "schemes": [
        "http,",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "title": "Sila API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/exercises": {
            "get": {
                "description": "Get a list of all exercises with information on muscle groups in the requested language (English or Russian). Supports filtering by muscle_group_id and searching by name.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Exercises"
                ],
                "summary": "Get all exercises",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Filter by muscle group ID",
                        "name": "muscle_group_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search by exercise name (both in Russian and English)",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of exercises",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ExerciseResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve exercises",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new exercise with names in both English and Russian, and specify the muscle group.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Exercises"
                ],
                "summary": "Create a new exercise",
                "parameters": [
                    {
                        "description": "Create new exercise",
                        "name": "exercise",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateExerciseRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Created exercise",
                        "schema": {
                            "$ref": "#/definitions/models.Exercise"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to create exercise",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/exercises/{id}": {
            "put": {
                "description": "Update an existing exercise with new names in both English and Russian, and update the muscle group.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Exercises"
                ],
                "summary": "Update an existing exercise",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Exercise ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update exercise",
                        "name": "exercise",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateExerciseRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated exercise",
                        "schema": {
                            "$ref": "#/definitions/models.Exercise"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Exercise not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to update exercise",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/muscle-groups": {
            "get": {
                "description": "Get a list of all muscle groups in the requested language (English or Russian).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MuscleGroups"
                ],
                "summary": "Get all muscle groups",
                "responses": {
                    "200": {
                        "description": "List of muscle groups",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.GetMuscleGroupsResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve muscle groups",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new muscle group with names in both English and Russian.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MuscleGroups"
                ],
                "summary": "Create a new muscle group",
                "parameters": [
                    {
                        "description": "Muscle group data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateMuscleGroupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Created muscle group",
                        "schema": {
                            "$ref": "#/definitions/models.MuscleGroup"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Failed to create muscle group",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/muscle-groups/{id}": {
            "put": {
                "description": "Обновление информации о группе мышц с учетом названий на двух языках.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MuscleGroups"
                ],
                "summary": "Update a muscle group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Muscle group ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated muscle group data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.MuscleGroup"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated muscle group",
                        "schema": {
                            "$ref": "#/definitions/models.MuscleGroup"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Muscle group not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Failed to update muscle group",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.CreateExerciseRequest": {
            "type": "object",
            "required": [
                "muscle_group_id",
                "name_eng",
                "name_ru"
            ],
            "properties": {
                "description_eng": {
                    "type": "string",
                    "example": "A basic chest exercise."
                },
                "description_ru": {
                    "type": "string",
                    "example": "Базовое упражнение для груди."
                },
                "image_url": {
                    "type": "string",
                    "example": "https://example.com/benchpress.jpg"
                },
                "muscle_group_id": {
                    "type": "integer",
                    "example": 1
                },
                "name_eng": {
                    "type": "string",
                    "example": "Bench Press"
                },
                "name_ru": {
                    "type": "string",
                    "example": "Жим лежа"
                }
            }
        },
        "models.CreateMuscleGroupRequest": {
            "type": "object",
            "required": [
                "name_eng",
                "name_ru"
            ],
            "properties": {
                "image_url": {
                    "type": "string",
                    "example": "http://example.com/chest.jpg"
                },
                "name_eng": {
                    "type": "string",
                    "example": "Chest"
                },
                "name_ru": {
                    "type": "string",
                    "example": "Грудь"
                }
            }
        },
...
        "models.Exercise": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description_eng": {
                    "type": "string"
                },
                "description_ru": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "muscle_group": {
                    "$ref": "#/definitions/models.MuscleGroup"
                },
                "name_eng": {
                    "type": "string"
                },
                "name_ru": {
                    "type": "string"
                }
            }
        },
        "models.ExerciseResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "muscle_group": {
                    "$ref": "#/definitions/models.MuscleGroupDTO"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "models.GetMuscleGroupsResponse": {
            "type": "object",
            "required": [
                "id",
                "name"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "models.MuscleGroup": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "name_eng": {
                    "type": "string"
                },
                "name_ru": {
                    "type": "string"
                }
            }
        },
        "models.MuscleGroupDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "models.UpdateExerciseRequest": {
            "type": "object",
            "required": [
                "muscle_group_id",
                "name_eng",
                "name_ru"
            ],
            "properties": {
                "description_eng": {
                    "type": "string",
                    "example": "A basic chest exercise."
                },
                "description_ru": {
                    "type": "string",
                    "example": "Базовое упражнение для груди."
                },
                "image_url": {
                    "type": "string",
                    "example": "https://example.com/benchpress.jpg"
                },
                "muscle_group_id": {
                    "type": "integer",
                    "example": 1
                },
                "name_eng": {
                    "type": "string",
                    "example": "Bench Press"
                },
                "name_ru": {
                    "type": "string",
                    "example": "Жим лежа"
                }
            }
        },
...
    }
}